+++ .ci/travis/install.sh
@@ -39,7 +39,7 @@ if [[ "$(uname -s)" == 'Darwin' ]]; then
 fi
 
 if [[ $TRAVIS_PYTHON_VERSION == '2.6' ]] || [[ $PYVER == 'py26' ]]; then
-    pip install -U ipaddress unittest2
+    pip install -U unittest2
 fi
 
-pip install -U pyopenssl pysendfile flake8
+pip install -U pyopenssl pysendfile flake8 mock